skills/contract-auditor/requirements.txt
# 合同审计 Skill 依赖
# 安装: pip install -r requirements.txt

# Word 文档处理
python-docx>=1.1.0

# PDF 处理（v1.1 启用）
# pdfplumber>=0.10.0

# OCR 识别（v1.1 启用）
# pytesseract>=0.3.10
# Pillow>=10.0.0

# Excel 报告（v1.2 启用）
# openpyxl>=3.1.0
